[navigation] move "login" and posts/adding components to separate screens
